api/requirements.txt
APIFlask==1.3.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
marshmallow-sqlalchemy==0.29.0